ai_service/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
easyocr
pillow
numpy
# opencv-python-headless is usually needed for easyocr if no GUI
opencv-python-headless
